sounds are routed to mixer channels
