--- Assets/menu_music.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b97ddeba25f3442b4aa2c55039812452
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
